-print0 was messing things up
also edit the icon generation script

+++ Icons/make_icons.sh
@@ -1,4 +1,8 @@
-mkdir PlatypusAppIcon.iconset
+#!/bin/bash
+
+if test ! -d PlatypusAppIcon.iconset; then
+  mkdir PlatypusAppIcon.iconset
+fi
 sips -z 16 16     PlatypusAppIcon1024.png --out PlatypusAppIcon.iconset/icon_16x16.png
 sips -z 32 32     PlatypusAppIcon1024.png --out PlatypusAppIcon.iconset/icon_16x16@2x.png
 sips -z 32 32     PlatypusAppIcon1024.png --out PlatypusAppIcon.iconset/icon_32x32.png
@@ -8,12 +12,19 @@ sips -z 256 256   PlatypusAppIcon1024.png --out PlatypusAppIcon.iconset/icon_128
 sips -z 256 256   PlatypusAppIcon1024.png --out PlatypusAppIcon.iconset/icon_256x256.png
 sips -z 512 512   PlatypusAppIcon1024.png --out PlatypusAppIcon.iconset/icon_256x256@2x.png
 sips -z 512 512   PlatypusAppIcon1024.png --out PlatypusAppIcon.iconset/icon_512x512.png
-cp PlatypusAppIcon1024.png PlatypusAppIcon.iconset/icon_512x512@2x.png
-#iconutil -c icns PlatypusAppIcon.iconset
-#rm -R PlatypusAppIcon.iconset
+if test -e PlatypusAppIcon1024.png; then
+  cp PlatypusAppIcon1024.png PlatypusAppIcon.iconset/icon_512x512@2x.png
+fi
+if test -x "$(which iconutil)" -a ! -e PlatypusAppIcon.iconset; then
+  iconutil -c icns PlatypusAppIcon.iconset
+  if test -d PlatypusAppIcon.iconset -a -w PlatypusAppIcon.iconset; then
+    touch PlatypusAppIcon.iconset || rm -R PlatypusAppIcon.iconset
+  fi
+fi
 
-
-mkdir PlatypusDefault.iconset
+if test ! -d PlatypusDefault.iconset; then
+  mkdir PlatypusDefault.iconset
+fi
 sips -z 16 16     PlatypusDefault1024.png --out PlatypusDefault.iconset/icon_16x16.png
 sips -z 32 32     PlatypusDefault1024.png --out PlatypusDefault.iconset/icon_16x16@2x.png
 sips -z 32 32     PlatypusDefault1024.png --out PlatypusDefault.iconset/icon_32x32.png
@@ -23,6 +34,12 @@ sips -z 256 256   PlatypusDefault1024.png --out PlatypusDefault.iconset/icon_128
 sips -z 256 256   PlatypusDefault1024.png --out PlatypusDefault.iconset/icon_256x256.png
 sips -z 512 512   PlatypusDefault1024.png --out PlatypusDefault.iconset/icon_256x256@2x.png
 sips -z 512 512   PlatypusDefault1024.png --out PlatypusDefault.iconset/icon_512x512.png
-cp PlatypusDefault1024.png PlatypusDefault.iconset/icon_512x512@2x.png
-#iconutil -c icns PlatypusDefault.iconset
-#rm -R PlatypusDefault.iconset
+if test -e PlatypusDefault1024.png; then
+  cp PlatypusDefault1024.png PlatypusDefault.iconset/icon_512x512@2x.png
+fi
+if test -x "$(which iconutil)" -a ! -e PlatypusDefault.icns; then
+  iconutil -c icns PlatypusDefault.iconset
+  if test -d PlatypusDefault.iconset -a -w PlatypusDefault.iconset; then
+    touch PlatypusDefault.iconset || rm -R PlatypusDefault.iconset
+  fi
+fi
